+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.2] - 2026-03-19
+
+### Fixed
+
+ - Исправлена ошибка вызова сервиса `mygas.get_bill` без указания даты: вместо первого дня текущего месяца (который мог отсутствовать в API) теперь используется дата последнего расчетного периода из `balances[0]["date"]` ([#20](https://github.com/lizardsystems/hass-mygas/issues/20)). Если данные о балансах отсутствуют, выполняется откат на первый день предыдущего месяца.
+
+## [2.1.1] - 2026-02-22
+
+### Fixed
+
+ - Исправлена ошибка, при которой события `mygas_*_failed` не отправлялись для исключений `HomeAssistantError`.
+
 ## [2.1.0] - 2026-02-22
 
 ### Added